+++ inst/embed/ui.R
@@ -2,8 +2,8 @@ library(shiny)
 library(googleAuthR)
 
 ## ui.R
-ui <- fluidPage(title = "googleAuthR Shiny Demo",
-  googleAuthUI("loginButton"),
+ui <- fluidPage(
+  gar_auth_jsUI("js_token"),
   textInput("url", "Enter URL"),
   actionButton("submit", "Shorten URL"),
   textOutput("short_url")